XQC sucking a banana
After the griefing of the Turkish flag by XQC, Papaplette and Mizkif, the turkish streamer Elraenn started his stream and fixed the Turkish flag. Turks driven by rage, wanted to embarrass XQC by drawing a portait of him sucking a banana. The portait was painted in less than a minute.